Minor improvements
