+++ test/RazorpayClientTest40.csproj
@@ -33,6 +33,9 @@
     <Reference Include="NUnit">
       <HintPath>..\packages\NUnit.3.6.1\lib\net40\nunit.framework.dll</HintPath>
     </Reference>
+    <Reference Include="Newtonsoft.Json">
+      <HintPath>..\packages\Newtonsoft.Json.7.0.1\lib\net45\Newtonsoft.Json.dll</HintPath>
+    </Reference>
     <Compile Include="Program.cs" />
     <Compile Include="Helper.cs" />
     <Compile Include="ClientTestCases.cs" />
